refactor(revenue-router): one update_config, accounts-based initialize, u64 counters
Review follow-ups on the router program:

- Replace set_admin, set_cranker, set_treasury and set_tiers with a single
  update_config(UpdateConfigArgs) whose fields are all Option. None leaves a
  field alone. The tier period lock moves inside the tiers branch.
- initialize takes admin, cranker and treasury as accounts instead of args and
  checks them with account constraints. The handler writes the config with
  set_inner so a new field cannot be skipped.
- distribute locks the cranker with an address constraint. The unused admin
  fallback is gone; the admin can point cranker at itself via update_config.
- The four stored counters and the two event fields become u64. Token amounts
  never exceed u64, and only the ladder math needs u128 for slice * bps.
- RouterConfig::pool_share wraps the ladder math for the handler.

IDL and SDK types regenerated. svm-tests cover update_config leaving unset
fields alone and initialize rejecting an aliasing treasury.

+++ programs/protocol-revenue-router/src/events.rs
@@ -29,6 +29,6 @@ pub struct FeesDistributed {
     pub to_treasury: u64,
     pub cap_room_after: u64,
     pub period_day: i64,
-    pub period_fees_after: u128,
-    pub lifetime_fees_after: u128,
+    pub period_fees_after: u64,
+    pub lifetime_fees_after: u64,
 }
